refactor(test): move testFactories to utils/test-utils and update imports
Move testFactories.ts and its test file from db/utils to utils/test-utils
to make them available for all renderer tests. Update import paths in db
repository tests to use the new location.
